requirement.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.11.0=pyhd3eb1b0_1
addict=2.4.0=pypi_0
aiohttp=3.6.3=py37h7b6447c_0
argon2-cffi=20.1.0=py37h27cfd23_1
async-timeout=3.0.1=py37h06a4308_0
async_generator=1.10=py37h28b3542_0
attrs=20.3.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=3.3.0=pyhd3eb1b0_0
blinker=1.4=py37h06a4308_0
blosc=1.21.0=h8c45485_0
brotli=1.0.9=he6710b0_2
brotlipy=0.7.0=py37h27cfd23_1003
brunsli=0.1=h2531618_0
bzip2=1.0.8=h7b6447c_0
c-ares=1.17.1=h27cfd23_0
ca-certificates=2021.1.19=h06a4308_0
cachetools=4.2.1=pyhd3eb1b0_0
cairo=1.14.12=h8948797_3
certifi=2020.12.5=py37h06a4308_0
cffi=1.14.5=py37h261ae71_0
chardet=3.0.4=py37h06a4308_1003
charls=2.1.0=he6710b0_2
click=7.1.2=pyhd3eb1b0_0
cloudpickle=1.6.0=py_0
coverage=5.5=py37h27cfd23_2
cryptography=3.3.1=py37h3c74f83_1
cudatoolkit=11.0.221=h6bb024c_0
cycler=0.10.0=py37_0
cython=0.29.22=py37h2531618_0
cytoolz=0.11.0=py37h7b6447c_0
dask-core=2021.2.0=pyhd3eb1b0_0
dataclasses=0.6=pypi_0
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=pyhd3eb1b0_0
defusedxml=0.7.0=pyhd3eb1b0_0
dominate=2.4.0=py_0
einops=0.3.0=pypi_0
entrypoints=0.3=py37_0
expat=2.2.10=he6710b0_2
ffmpeg=4.0=hcdf2ecd_0
fontconfig=2.13.1=h6c09931_0
freeglut=3.0.0=hf484d3e_5
freetype=2.10.4=h5ab3b9f_0
future=0.18.2=pypi_0
fvcore=0.1.2.post20201204=pyhd8ed1ab_0
giflib=5.1.4=h14c3975_1
glib=2.67.4=h36276a3_1
google-auth=1.27.1=pyhd3eb1b0_0
google-auth-oauthlib=0.4.3=pyhd3eb1b0_0
graphite2=1.3.14=h23475e2_0
grpcio=1.36.1=py37h2157cd5_1
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
icu=58.2=he6710b0_3
idna=2.10=pyhd3eb1b0_0
imagecodecs=2021.1.11=py37h581e88b_1
imageio=2.9.0=py_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
intel-openmp=2020.2=254
ipykernel=5.3.4=py37h5ca1d4c_0
ipython=7.21.0=py37hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.3=pyhd3eb1b0_1
jasper=2.0.14=h07fcdf6_1
jedi=0.17.0=py37_0
jinja2=2.11.3=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
jsonpatch=1.28=pypi_0
jsonpointer=2.0=pypi_0
jsonschema=3.2.0=py_2
jupyter=1.0.0=py37_7
jupyter_client=6.1.7=py_0
jupyter_console=6.2.0=py_0
jupyter_core=4.7.1=py37h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
jxrlib=1.1=h7b6447c_2
kaggle=1.5.10=pypi_0
kiwisolver=1.3.1=py37h2531618_0
kornia=0.4.1=pypi_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
lerc=2.2.1=h2531618_0
libaec=1.0.4=he6710b0_1
libdeflate=1.7=h27cfd23_5
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libglu=9.0.0=hf484d3e_1
libopencv=3.4.2=hb342d67_1
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.14.0=h8c45485_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libwebp=1.0.1=h8e7db2f_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
libzopfli=1.0.3=he6710b0_0
lpips=0.1.3=pypi_0
lz4-c=1.9.3=h2531618_0
markdown=3.3.4=py37h06a4308_0
markupsafe=1.1.1=py37h14c3975_1
matplotlib=3.3.2=0
matplotlib-base=3.3.2=py37h817c723_0
mistune=0.8.4=py37h14c3975_1001
mkl=2020.2=256
mkl-service=2.3.0=py37he8ac12f_0
mkl_fft=1.3.0=py37h54f3939_0
mkl_random=1.1.1=py37h0573a6f_0
mmaction2=0.12.0=dev_0
mmcv-full=1.3.0=pypi_0
multidict=4.7.6=py37h7b6447c_1
multiscaledeformableattention=1.0=pypi_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.0.7=py37_0
nbformat=5.1.2=pyhd3eb1b0_1
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
networkx=2.5=py_0
ninja=1.10.2=py37hff7bd54_0
notebook=6.2.0=py37h06a4308_0
numpy=1.19.2=py37h54aff64_0
numpy-base=1.19.2=py37hfa32c7d_0
oauthlib=3.1.0=py_0
olefile=0.46=py37_0
opencv=3.4.2=py37h6fd60c2_1
opencv-contrib-python=4.5.1.48=pypi_0
opencv-python=4.5.1.48=pypi_0
openjpeg=2.3.0=h05c96fa_1
openssl=1.1.1j=h27cfd23_0
packaging=20.8=pypi_0
pandas=1.2.3=py37ha9443f7_0
pandoc=2.11=hb0f4dca_0
pandocfilters=1.4.3=py37h06a4308_1
parso=0.8.1=pyhd3eb1b0_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.1.1=py37he98fc37_0
pip=21.0.1=py37h06a4308_0
pixman=0.40.0=h7b6447c_0
portalocker=2.2.1=py37h06a4308_0
prometheus_client=0.9.0=pyhd3eb1b0_0
prompt-toolkit=3.0.8=py_0
prompt_toolkit=3.0.8=0
protobuf=3.14.0=py37h2531618_1
psutil=5.7.2=py37h7b6447c_0
ptyprocess=0.7.0=pyhd3eb1b0_2
py-opencv=3.4.2=py37hb342d67_1
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycocotools=2.0.2=pypi_0
pycparser=2.20=py_2
pygments=2.8.0=pyhd3eb1b0_0
pyjwt=1.7.1=py37_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py37h05f1152_2
pyrsistent=0.17.3=py37h7b6447c_0
pysocks=1.7.1=py37_1
python=3.7.9=h7579374_0
python-dateutil=2.8.1=pyhd3eb1b0_0
python-slugify=4.0.1=pypi_0
pytoml=0.1.21=pypi_0
pytorch=1.7.0=py3.7_cuda11.0.221_cudnn8.0.3_0
pytorch-fid=0.2.1=pypi_0
pytz=2021.1=pyhd3eb1b0_0
pywavelets=1.1.1=py37h7b6447c_2
pyyaml=5.3.1=py37h7b6447c_1
pyzmq=20.0.0=pypi_0
qt=5.9.7=h5867ecd_1
qtconsole=5.0.2=pyhd3eb1b0_0
qtpy=1.9.0=py_0
rawpy=0.16.0=pypi_0
readline=8.1=h27cfd23_0
requests=2.25.0=pypi_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-image=0.17.2=py37hdf5156a_0
scipy=1.6.1=py37h91f5cce_0
seaborn=0.11.1=pyhd3eb1b0_0
send2trash=1.5.0=pyhd3eb1b0_1
setproctitle=1.2.2=py37h27cfd23_1004
setuptools=52.0.0=py37h06a4308_0
sip=4.19.8=py37hf484d3e_0
six=1.15.0=py37h06a4308_0
snappy=1.1.8=he6710b0_0
sqlite=3.33.0=h62c20be_0
tabulate=0.8.9=py37h06a4308_0
tensorboard=2.4.0=pyhc547734_0
tensorboard-plugin-wit=1.6.0=py_0
tensorboardx=2.2=pypi_0
termcolor=1.1.0=py37h06a4308_1
terminado=0.9.2=py37h06a4308_0
testpath=0.4.4=pyhd3eb1b0_0
text-unidecode=1.3=pypi_0
tifffile=2021.3.4=pyhd3eb1b0_1
timm=0.4.5=pypi_0
tk=8.6.10=hbc83047_0
toolz=0.11.1=pyhd3eb1b0_0
torchaudio=0.7.0=py37
torchfile=0.1.0=pypi_0
torchvision=0.8.1=py37_cu110
tornado=6.1=py37h27cfd23_0
tqdm=4.50.2=py_0
traitlets=5.0.5=pyhd3eb1b0_0
typing_extensions=3.7.4.3=pyha847dfd_0
urllib3=1.26.2=pypi_0
visdom=0.1.8.9=pypi_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py37_1
websocket-client=0.57.0=pypi_0
werkzeug=1.0.1=pyhd3eb1b0_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=py37_0
xz=5.2.5=h7b6447c_0
yacs=0.1.6=py_0
yaml=0.2.5=h7b6447c_0
yapf=0.31.0=pypi_0
yarl=1.6.3=py37h27cfd23_0
zeromq=4.3.3=he6710b0_3
zfp=0.5.5=h2531618_4
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
